services/study-engine/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.24.1
structlog==24.1.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.5
pytest-httpx==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
PyJWT==2.8.0
